+++ packages/plugin/src/theia.d.ts
@@ -6844,8 +6844,7 @@ export module '@theia/plugin' {
      * thenable.
      *
      */
-    export type ProviderResult<T> = T | undefined | PromiseLike<T | undefined>;
-
+    export type ProviderResult<T> = T | undefined | null | Thenable<T | undefined | null>;
     /**
      * A symbol kind.
      */
@@ -9662,10 +9661,7 @@ export module '@theia/plugin' {
          * @param token A cancellation token.
          * @return An array of {@link DebugConfiguration debug configurations}.
          */
-        provideDebugConfigurations?(
-            folder: WorkspaceFolder | undefined,
-            token?: CancellationToken
-        ): ProviderResult<DebugConfiguration[]>;
+        provideDebugConfigurations?(folder: WorkspaceFolder | undefined, token?: CancellationToken): ProviderResult<DebugConfiguration[]>;
 
         /**
          * Resolves a [debug configuration](#DebugConfiguration) by filling in missing values or by adding/changing/removing attributes.
@@ -9679,11 +9675,7 @@ export module '@theia/plugin' {
          * @param token A cancellation token.
          * @return The resolved debug configuration or undefined or null.
          */
-        resolveDebugConfiguration?(
-            folder: WorkspaceFolder | undefined,
-            debugConfiguration: DebugConfiguration,
-            token?: CancellationToken
-        ): ProviderResult<DebugConfiguration | undefined | null>;
+        resolveDebugConfiguration?(folder: WorkspaceFolder | undefined, debugConfiguration: DebugConfiguration, token?: CancellationToken): ProviderResult<DebugConfiguration>;
 
         /**
          * This hook is directly called after 'resolveDebugConfiguration' but with all variables substituted.
@@ -9698,11 +9690,7 @@ export module '@theia/plugin' {
          * @param token A cancellation token.
          * @return The resolved debug configuration or undefined or null.
          */
-        resolveDebugConfigurationWithSubstitutedVariables?(
-            folder: WorkspaceFolder | undefined,
-            debugConfiguration: DebugConfiguration,
-            token?: CancellationToken
-        ): ProviderResult<DebugConfiguration | undefined | null>;
+        resolveDebugConfigurationWithSubstitutedVariables?(folder: WorkspaceFolder | undefined, debugConfiguration: DebugConfiguration, token?: CancellationToken): ProviderResult<DebugConfiguration>;
     }
 
     /**